test: fix & add benchmark into CI build
